requirements.txt
scikit-surprise                    
scikit-learn==1.2.2
pandas
 
